requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0
motor
redis
aioredis
python-jose[cryptography]
passlib[bcrypt]
redis[asyncio]
